Usability improvements (itemized below)1. DataContext will not auto-create itself if no config exists; use DataContext.create() for a new instance.2. ge.validate will now accept a path to a data_context for the data_context parameter instead of requiring an instantiated object3. Clearer error handling for above situations4. Slack will use uuidv1 instead of v4 to autogenerate random names5. Refactoring to support above changes6. Changes to associated tests
