+++ great_expectations/cli/cli.py
@@ -253,6 +253,8 @@ To launch with jupyter notebooks:
 To launch with jupyter lab: 
     jupyter lab great_expectations/notebooks/create_expectations_for_spark_dataframes.ipynb
     """
+    context = DataContext.create('.')
+
     base_dir = os.path.join(target_directory, "great_expectations")
 
     cli_message("Great Expectations", color="cyan", figlet=True)
@@ -269,7 +271,6 @@ To launch with jupyter lab:
         "\nDone.",
         color="blue")
 
-    context = DataContext('.')
 
     # Shows a list of options to select from
 
@@ -291,7 +292,9 @@ To launch with jupyter lab:
         if path.startswith("./"):
             path = path[2:]
 
-        default_data_source_name = os.path.basename(path)
+        if path.endswith("/"):
+            basenamepath = path[:-1]
+        default_data_source_name = os.path.basename(basenamepath)
         data_source_name = click.prompt(
             msg_prompt_datasource_name, default=default_data_source_name, show_default=True)
