+++ tests/conftest.py
@@ -8,7 +8,7 @@ import numpy as np
 import sqlalchemy as sa
 
 import great_expectations as ge
-from great_expectations.util import safe_mmkdir
+from great_expectations.data_context.util import safe_mmkdir
 
 from .test_utils import get_dataset
 
@@ -107,12 +107,14 @@ def sqlitedb_engine():
 def empty_data_context(tmp_path_factory):
     context_path = tmp_path_factory.mktemp('empty_data_context')
     context_path = str(context_path)
+    context = ge.data_context.DataContext.create(context_path)
     asset_config_path = os.path.join(context_path, "great_expectations/expectations")
     safe_mmkdir(asset_config_path, exist_ok=True)
-    return ge.data_context.DataContext(context_path)
+    return context
 
 @pytest.fixture()
 def data_context(tmp_path_factory):
+    # This data_context is *manually* created to have the config we want, vs created with DataContext.create
     context_path = tmp_path_factory.mktemp('data_context')
     context_path = str(context_path)
     asset_config_path = os.path.join(context_path, "great_expectations/expectations")
