+++ tests/data_context/test_data_context.py
@@ -8,7 +8,7 @@ import sqlalchemy as sa
 import pandas as pd
 
 from great_expectations.data_context import DataContext
-from great_expectations.util import safe_mmkdir
+from great_expectations.data_context.util import safe_mmkdir
 # get_data_context
 from great_expectations.dataset import PandasDataset, SqlAlchemyDataset
 
@@ -40,23 +40,6 @@ def parameterized_expectations_config():
         ]
     }
 
-@pytest.fixture()
-def parameterized_config_data_context(tmp_path_factory):
-    context_path = tmp_path_factory.mktemp("empty_context_dir")
-    context_path = str(context_path)
-    asset_config_path = os.path.join(context_path, "great_expectations/expectations")
-    safe_mmkdir(asset_config_path, exist_ok=True)
-    shutil.copy("./tests/test_fixtures/great_expectations_basic.yml", str(context_path))
-    shutil.copy("./tests/test_fixtures/expectations/parameterized_expectations_config_fixture.json",str(asset_config_path))
-    return DataContext(context_path)
-
-
-# def test_invalid_data_context():
-#     # Test an unknown data context name
-#     with pytest.raises(ValueError) as err:
-#         get_data_context('what_a_ridiculous_name', None)
-#         assert "Unknown data context." in str(err)
-
 def test_list_available_data_asset_names(empty_data_context, filesystem_csv):
     empty_data_context.add_datasource("my_datasource", "pandas", base_directory= str(filesystem_csv))
     available_asset_names = empty_data_context.list_available_data_asset_names() 
@@ -68,23 +51,22 @@ def test_list_available_data_asset_names(empty_data_context, filesystem_csv):
             "available_data_asset_names": set(["f1", "f2", "f3"])
         }]
     }]
-    # assert data_context.list_available_data_asset_names() == ['parameterized_expectations_config_fixture']
 
 def test_list_expectations_configs(data_context):
     assert data_context.list_expectations_configs() == ['parameterized_expectations_config_fixture']
 
-def test_get_existing_data_asset_config(parameterized_config_data_context):
-    data_asset_config = parameterized_config_data_context.get_expectations('parameterized_expectations_config_fixture')
+def test_get_existing_data_asset_config(data_context):
+    data_asset_config = data_context.get_expectations('parameterized_expectations_config_fixture')
     assert data_asset_config['data_asset_name'] == 'parameterized_expectations_config_fixture'
     assert len(data_asset_config['expectations']) == 2
 
-def test_get_new_data_asset_config(parameterized_config_data_context):
-    data_asset_config = parameterized_config_data_context.get_expectations('this_data_asset_config_does_not_exist')
+def test_get_new_data_asset_config(data_context):
+    data_asset_config = data_context.get_expectations('this_data_asset_config_does_not_exist')
     assert data_asset_config['data_asset_name'] == 'this_data_asset_config_does_not_exist'
     assert len(data_asset_config['expectations']) == 0
 
-def test_save_data_asset_config(parameterized_config_data_context):
-    data_asset_config = parameterized_config_data_context.get_expectations('this_data_asset_config_does_not_exist')
+def test_save_data_asset_config(data_context):
+    data_asset_config = data_context.get_expectations('this_data_asset_config_does_not_exist')
     assert data_asset_config['data_asset_name'] == 'this_data_asset_config_does_not_exist'
     assert len(data_asset_config['expectations']) == 0
     data_asset_config['expectations'].append({
@@ -93,11 +75,11 @@ def test_save_data_asset_config(parameterized_config_data_context):
                 "value": 10
             }
         })
-    parameterized_config_data_context.save_expectations(data_asset_config)
-    data_asset_config_saved = parameterized_config_data_context.get_expectations('this_data_asset_config_does_not_exist')
+    data_context.save_expectations(data_asset_config)
+    data_asset_config_saved = data_context.get_expectations('this_data_asset_config_does_not_exist')
     assert data_asset_config['expectations'] == data_asset_config_saved['expectations']
 
-def test_register_validation_results(parameterized_config_data_context):
+def test_register_validation_results(data_context):
     run_id = "460d61be-7266-11e9-8848-1681be663d3e"
     source_patient_data_results = {
         "meta": {"data_asset_name": "source_patient_data"},
@@ -123,9 +105,9 @@ def test_register_validation_results(parameterized_config_data_context):
         ],
         "success": True
     }
-    res = parameterized_config_data_context.register_validation_results(run_id, source_patient_data_results)
+    res = data_context.register_validation_results(run_id, source_patient_data_results)
     assert res == source_patient_data_results # results should always be returned, and in this case not modified
-    bound_parameters = parameterized_config_data_context._evaluation_parameter_store.get_run_parameters(run_id)
+    bound_parameters = data_context._evaluation_parameter_store.get_run_parameters(run_id)
     assert bound_parameters == {
         'urn:great_expectations:validations:source_patient_data:expectations:expect_table_row_count_to_equal:result:observed_value': 1024
     }
@@ -155,16 +137,16 @@ def test_register_validation_results(parameterized_config_data_context):
         ],
         "success": True
     }
-    parameterized_config_data_context.register_validation_results(run_id, source_diabetes_data_results)
-    bound_parameters = parameterized_config_data_context._evaluation_parameter_store.get_run_parameters(run_id)
+    data_context.register_validation_results(run_id, source_diabetes_data_results)
+    bound_parameters = data_context._evaluation_parameter_store.get_run_parameters(run_id)
     assert bound_parameters == {
         'urn:great_expectations:validations:source_patient_data:expectations:expect_table_row_count_to_equal:result:observed_value': 1024, 
         'urn:great_expectations:validations:source_diabetes_data:expectations:expect_column_unique_value_count_to_be_between:columns:patient_nbr:result:observed_value': 2048
     }
 
-def test_compile(parameterized_config_data_context):
-    parameterized_config_data_context._compile()
-    assert parameterized_config_data_context._compiled_parameters == {
+def test_compile(data_context):
+    data_context._compile()
+    assert data_context._compiled_parameters == {
         'raw': {
             'urn:great_expectations:validations:source_diabetes_data:expectations:expect_column_unique_value_count_to_be_between:columns:patient_nbr:result:observed_value', 
             'urn:great_expectations:validations:source_patient_data:expectations:expect_table_row_count_to_equal:result:observed_value'
@@ -194,14 +176,11 @@ def test_compile(parameterized_config_data_context):
 def test_normalize_data_asset_names(tmp_path_factory):
     base_dir = tmp_path_factory.mktemp("test_normalize_data_asset_names")
     base_dir = str(base_dir)
-    os.makedirs(os.path.join(base_dir, "great_expectations"))
-    # context_dir = os.path.join(base_dir, "great_expectations")
+    context = DataContext.create(base_dir)
     # asset_dir = context_dir.join("expectations/ds1/gen1/data_asset_1/")
     # os.makedirs(asset_dir)
     # with open(asset_dir("default.json"), "w") as config:
     #     json.dump({"data_asset_name": "data_assset_1"}, config)
-
-    context = DataContext(base_dir)
 
     # assert context._normalize_data_asset_name("data_asset_1") == "ds1/gen1/data_asset_1"
     # NOTE: NORMALIZATION IS CURRENTLY A NO-OP
